:sparkles: 신청여부 예외처리 구현

+++ src/main/java/Idea/Archive/IdeaArchive/domain/application/service/CancelApplicationService.java
@@ -1,5 +1,6 @@
 package Idea.Archive.IdeaArchive.domain.application.service;
 
+import Idea.Archive.IdeaArchive.domain.application.exception.NotApplyApplicationException;
 import Idea.Archive.IdeaArchive.domain.application.repository.ApplicationRepository;
 import Idea.Archive.IdeaArchive.domain.member.entity.Member;
 import Idea.Archive.IdeaArchive.domain.post.entity.Post;
@@ -23,6 +24,9 @@ public class CancelApplicationService {
         Member member = memberUtil.currentMember();
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new NotExistPostException("게시글이 존재하지 않습니다"));
+        if(!applicationRepository.existsByPostAndMember(post,member)){
+            throw new NotApplyApplicationException("신청하지 않았습니다");
+        }
         post.updateApplication(post.getApplicationCount()-1);
         postRepository.save(post);
         applicationRepository.deleteByPostAndMember(post,member);
